myvestige-fastapi/app/dependencies/auth.py
from fastapi import Depends, HTTPException, Security
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from app.core.config import settings
from app.core.security import decode_access_token
from app.models.user import User  # Assuming you have a User model
from app.db.session import get_db
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy.future import select

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")


async def get_current_user(token: str = Security(oauth2_scheme), db: AsyncSession = Depends(get_db)):
    """Extract user from JWT token"""
    credentials_exception = HTTPException(status_code=401, detail="Invalid authentication credentials")

    payload = decode_access_token(token)
    if payload is None:
        raise credentials_exception

    user_id: str = payload.get("sub")
    if user_id is None:
        raise credentials_exception

    # Query user from DB
    result = await db.execute(select(User).where(User.id == user_id))
    user = result.scalars().first()

    if not user:
        raise credentials_exception
    return user