myvestige-fastapi/requirements/base.txt
fastapi==0.95.0
uvicorn==0.18.2
sqlalchemy==2.0.0
psycopg2-binary==2.9.3
alembic==1.9.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-admin==1.0.4